requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.0

# Optional dependencies for notifications
# Uncomment if you want to use SMS notifications
# twilio>=8.10.0

# Optional dependencies for enhanced features
python-dateutil>=2.8.2